pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["wathika-eng <wathika02@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
werkzeug = "^3.0.3"
faker = "^26.0.0"
drf-spectacular = "^0.27.2"
pip-review = "^1.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
